VER 1.1.0-beta.2
-- Add save command on close window event.
x# Update ".gitignore".
x- Fix some bugs (Clock).
x- Output Framework on start.
